winvnc: start reverting changes
Revert changes made to fix compiler warnings without any actual impact
in order to stay as close as possible to upstream.

+++ winvnc/winvnc/vncEncodeZlibHex.h
@@ -52,13 +52,14 @@ class vncEncodeZlibHex;
 #else
 #include "lzo/minilzo.h"
 #endif
+
 // Minimum zlib rectangle size in bytes.  Anything smaller will
 // not compress well due to overhead.
 // temp change lzo
 //#define VNC_ENCODE_ZLIBHEX_MIN_COMP_SIZE (17)
 #define VNC_ENCODE_ZLIBHEX_MIN_COMP_SIZE (64)
 // Flag used to mark our compressors as uninitialized.
-#define ZLIBHEX_COMP_UNINITED ((DWORD)-1)
+#define ZLIBHEX_COMP_UNINITED (-1)
 
 // Size of the smallest update portion sent independently across
 // the network.  This encoder can transmit partial updates to
